Creating a wrapper file that acts as a place to hold helpers and initializers for the rest of the items
